Move JWT tests into Authentication tests
